requirements.txt
opencv-python
ultralytics
segment_anything
huggingface-hub
diffusers @ git+https://github.com/huggingface/diffusers.git@8fcfb2a456e5c35d6d532faccf4859d303c22501
transformers @ git+https://github.com/huggingface/transformers.git@dc8b6eaeeeb59dd3089b478cc09b577f2c62a297
celery
fastapi
uvicorn
SQLAlchemy
python-multipart
accelerate
python-dotenv
manga-ocr
craft-text-detector
openai
scikit-learn
lxml
